additional files fist part

+++ app/Http/Controllers/TourController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers; 
 use App\Models\Tours; 
+use App\Models\MediaFiles; 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -62,27 +63,44 @@ class TourController extends Controller
 
     public function updateImg($id, Request $request) {
 
-        $tour = Tours::find($id); 
-
         $request->validate([
             'image' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
         ]);
 
         $imageName = time().'.'.$request->image->extension();  
-     
+
         $request->image->move(public_path('assets/uploads/tour-images/'), $imageName);
 
-        $oldPicture = $tour->picture;
+        if($request->fileType == "0") {
 
-        $tour->picture = $imageName; 
+            $tour = Tours::find($id); 
+  
+            $oldPicture = $tour->picture;
+    
+            $tour->picture = $imageName; 
+    
+            if(unlink(public_path('assets/uploads/tour-images/'.$oldPicture)))
+                if($tour->save()) {  
+                    return back()->with('success','Məlumat yeniləndi');
+                } else { 
+                    return back()->with('error','Xəta baş verdi, zəhmət olmasa biraz sora yenidən cəhd edin');
+                }
+        } else {
 
-        if(unlink(public_path('assets/uploads/tour-images/'.$oldPicture)))
-            if($tour->save()) { 
-                
-                return back()->with('success','Məlumat yeniləndi');
+            $media = new MediaFiles();
+
+            $media->section_id = $id;
+            $media->section_id = $id;
+            $media->type = 1;
+            $media->file = $imageName; 
+
+            if($media->save()) {  
+                return back()->with('success','Şəkil əlavə olundu');
             } else { 
                 return back()->with('error','Xəta baş verdi, zəhmət olmasa biraz sora yenidən cəhd edin');
             }
+        }
+        
 
     }
